backend/requirements.txt
ansicon==1.89.0
asgiref==3.7.2
astroid==2.15.5
asttokens==2.2.1
autopep8==2.0.2
backcall==0.2.0
backports.zoneinfo;python_version<"3.9"
black==23.3.0
blessed==1.20.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.1
curtsies==0.4.1
cwcwidth==0.1.8
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
Django==4.2.2
django-extensions==3.2.3
django-stubs==4.2.3
django-stubs-ext==4.2.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
exceptiongroup==1.1.1
executing==1.2.0
flake8==6.0.0
greenlet==2.0.2
idna==3.4
iniconfig==2.0.0
ipython==8.12.2
isort==5.12.0
jedi==0.18.2
jinxed==1.2.0
lazy-object-proxy==1.9.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy==1.4.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
parso==0.8.3
pathspec==0.11.1
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.8.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==2.7.0
pylint==2.17.4
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
pytest==5.4.0
pytest-django==4.5.2
pytest-pythonpath==0.7.4
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
pyxdg==0.28
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
stack-data==0.6.2
tomli==2.0.1
tomlkit==0.11.8
traitlets==5.9.0
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.10
typing-extensions==4.7.0
tzdata==2023.3
urllib3==2.0.3
wcwidth==0.2.6
wrapt==1.15.0
